refactor: split index.js code into 4 exercise components
